refactor: changes type files to .d.ts
