+++ frontend/app/pages/event/[eventId]/book/seats/(types)/event.d.ts
