+++ frontend/app/pages/event/[eventId]/book/seats/(types)/seating-layout.d.ts
