+++ frontend/app/pages/event/[eventId]/book/seats/(types)/ticket.d.ts
